Added benchmark!
