+++ package.json
@@ -22,6 +22,7 @@
   },
   "dependencies": {
     "@angular/core": "~2.4.7",
+    "@angular/forms": "^2.4.10",
     "@angular/router": "^3.4.8",
     "rxjs": "~5.0.3",
     "uwp-keycodes": "^1.0.1"
